Create another Shape to Graphic demo for familiarity
This demo was used to figure out how shapes are placed.
On my machine, the pixelsPerInch seems about right.

A rectangle is described with 2 points, which are it's total length.
It it is then centered over the origin.

An ellipse is described with 2 points, which half it's total lenght.
It is mirrored over the origin.

The right triangle is described with 2 points, which describe its
legs in the x and y directions, respectively.

The polygon is described with n points, which are it's vertices.
This was the simplest for me to understand with the coordinate
mapping fixed.