Add xor and determineOrientation functions
I have been puzzling for the past few days about how to determine if a list
of Coordinates creates a clockwise or counter-clockwise orientation of a
polygon.

I deferred to Stack Overflow, and found a relatively algorithm and added it
here. Still need to use that information in the definition of containsS for
a Polygon, but wanted to capture it.